requirements.txt
# Discord Bot
discord.py==2.3.2
aiohttp==3.9.1

# Google Sheets API
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0

# Utilities
python-dotenv==1.0.0
Pillow==10.1.0

# Async support
asyncio==3.4.3
